refactor: add python directory to accommodate Python things
